+++ model/bill.go
@@ -11,7 +11,7 @@ type BillType struct {
 // Bill keeps information about particular bill
 type Bill struct {
 	ID          int       `json:"id" db:"id"`
-	Price       float32   `json:"total_price" db:"total_price"`
+	Price       float64   `json:"total_price" db:"total_price"`
 	DueDate     time.Time `json:"due_date" db:"due_date"`
 	PeriodStart time.Time `json:"period_start" db:"period_start"`
 	PeriodEnd   time.Time `json:"period_end" db:"period_end"`
@@ -23,7 +23,7 @@ type Bill struct {
 type TenantBill struct {
 	ID          int        `json:"id" db:"id"`
 	TenantID    int        `json:"tenant_id" db:"tenant_id"`
-	Price       float32    `json:"price" db:"price"`
+	Price       float64    `json:"price" db:"price"`
 	PaymentDate *time.Time `json:"payment_date" db:"payment_date"`
 	Bill        `json:"bill" db:"bill"`
 }
